Moved iterator to new utility header.
